feat(settings): move the sleep timer dial behind a configure button
