+++ assets/i18n/pl/main.ftl
@@ -531,6 +531,10 @@ settings-normalisation = Normalizacja głośności
 settings-normalisation-detail = Utrzymuje stałą głośność utworów
 settings-gapless = Odtwarzanie bez przerw
 settings-gapless-detail = Przechodzi z utworu do następnego bez pauzy, zgodnie z układem albumu
+settings-sleep-configure = Skonfiguruj…
+settings-sleep-off = Wył
+settings-sleep-end-of-track = Na końcu utworu
+settings-sleep-minutes = { $count } min
 settings-panel-lyrics-size = Rozmiar tekstu utworu (panel)
 settings-panel-lyrics-size-detail = Rozmiar wierszy tekstu utworu w panelu bocznym ponad bazowy rozmiar czcionki
 settings-fullscreen-lyrics-size = Rozmiar tekstu utworu (pełny ekran)
