+++ assets/i18n/ru/main.ftl
@@ -446,6 +446,10 @@ settings-normalisation = Нормализация громкости
 settings-normalisation-detail = Держит треки на одинаковой громкости
 settings-gapless = Воспроизведение без пауз
 settings-gapless-detail = Переводит трек в следующий без паузы, как задумано в альбоме
+settings-sleep-configure = Настроить…
+settings-sleep-off = Выкл
+settings-sleep-end-of-track = В конце трека
+settings-sleep-minutes = { $count } мин
 settings-panel-lyrics-size = Размер текста песни (панель)
 settings-panel-lyrics-size-detail = Размер строк текста песни на боковой панели поверх базового размера шрифта
 settings-fullscreen-lyrics-size = Размер текста песни (полный экран)
