+++ assets/i18n/uk/main.ftl
@@ -446,6 +446,10 @@ settings-normalisation = Нормалізація гучності
 settings-normalisation-detail = Тримає треки на однаковій гучності
 settings-gapless = Відтворення без пауз
 settings-gapless-detail = Переводить трек у наступний без паузи, як задумано в альбомі
+settings-sleep-configure = Налаштувати…
+settings-sleep-off = Вимк
+settings-sleep-end-of-track = Наприкінці треку
+settings-sleep-minutes = { $count } хв
 settings-panel-lyrics-size = Розмір тексту пісні (панель)
 settings-panel-lyrics-size-detail = Розмір рядків тексту пісні на бічній панелі поверх базового розміру шрифту
 settings-fullscreen-lyrics-size = Розмір тексту пісні (повний екран)
